Image Caption Project/requirements.txt
fastapi
uvicorn
transformers
torch 
Pillow
streamlit
requests
python-multipart